+++ usr/src/sys/vm/vm_page.h
@@ -7,7 +7,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)vm_page.h	8.2 (Berkeley) %G%
+ *	@(#)vm_page.h	8.3 (Berkeley) %G%
  *
  *
  * Copyright (c) 1987, 1990 Carnegie-Mellon University.
@@ -176,14 +176,14 @@ simple_lock_data_t	vm_page_queue_free_lock;
 
 #define PAGE_ASSERT_WAIT(m, interruptible)	{ \
 				(m)->flags |= PG_WANTED; \
-				assert_wait((int) (m), (interruptible)); \
+				assert_wait((m), (interruptible)); \
 			}
 
 #define PAGE_WAKEUP(m)	{ \
 				(m)->flags &= ~PG_BUSY; \
 				if ((m)->flags & PG_WANTED) { \
 					(m)->flags &= ~PG_WANTED; \
-					thread_wakeup((int) (m)); \
+					thread_wakeup((m)); \
 				} \
 			}
